requirements.txt
click
colorama
mozjpeg_lossless_optimization
onnxruntime>=1.14.0
Pillow
opencv-python
numpy